Removed the deprecated json matchers
