+++ kotest-assertions/kotest-assertions-json/src/commonMain/kotlin/io/kotest/assertions/json/comparisons/CompareBooleans.kt
@@ -1,0 +1,10 @@
+package io.kotest.assertions.json.comparisons
+
+import io.kotest.assertions.json.JsonError
+
+internal fun compareBooleans(path: List<String>, expected: Boolean, actual: Boolean): JsonError? {
+   return when (expected) {
+      actual -> null
+      else -> JsonError.UnequalBooleans(path, expected, actual)
+   }
+}
