+++ kotest-assertions/kotest-assertions-json/src/commonMain/kotlin/io/kotest/assertions/json/comparisons/CompareStrings.kt
@@ -1,0 +1,10 @@
+package io.kotest.assertions.json.comparisons
+
+import io.kotest.assertions.json.JsonError
+
+internal fun compareStrings(path: List<String>, expected: String, actual: String): JsonError? {
+   return when (expected) {
+      actual -> null
+      else -> JsonError.UnequalStrings(path, expected, actual)
+   }
+}
